create ClientBase class
